add a fixture to conftest.py
